feat: move edit profile, delegations and treasuries into settings page (#653)
* fix: ceil scrollLeft

For some reason scrollBy might not scroll all the way (for example 199.5 instead of 200).
Further calls won't cause another scroll so we might be missing 0.5 to the right.

This changes it so if we are off by less than 1 pixel we still treat it as scrolled
all the way to the right.

* feat: move profile settings to Settings

* fix: block form on formErrors

* fix: make navigation sticky

* chore: update settings tab layout

* fix: clear formErrors on reset

* fix: truncate treasury and delegation names

* feat: make sure elements scroll into view on focus

* feat: make gradient step more visible on xxl

* fix: use v-show for profile tab to retain dirty flags

More general solution is needed: https://github.com/snapshot-labs/sx-monorepo/issues/573

* fix: update paddings

* fix: scroll to end

This somehow prevents vertical scroll that causes profile page to scroll
away.

* fix: assure padding on small screens

* fix: increase sticky menu z-index to 50
