+++ apps/ui/src/networks/starknet/actions.ts
@@ -565,6 +565,7 @@ export function createActions(
     updateSettings: async (
       web3: any,
       space: Space,
+      metadata: SpaceMetadata,
       authenticatorsToAdd: StrategyConfig[],
       authenticatorsToRemove: number[],
       votingStrategiesToAdd: StrategyConfig[],
@@ -574,6 +575,14 @@ export function createActions(
       minVotingDuration: number | null,
       maxVotingDuration: number | null
     ) => {
+      const pinned = await helpers.pin(
+        createErc1155Metadata(metadata, {
+          execution_strategies: space.executors,
+          execution_strategies_types: space.executors_types,
+          execution_destinations: space.executors_destinations
+        })
+      );
+
       const metadataUris = await Promise.all(
         votingStrategiesToAdd.map(config => buildMetadata(helpers, config))
       );
@@ -587,6 +596,7 @@ export function createActions(
         signer: web3.provider.account,
         space: space.id,
         settings: {
+          metadataUri: `ipfs://${pinned.cid}`,
           authenticatorsToAdd: authenticatorsToAdd.map(
             config => config.address
           ),
